JSP receitas

+++ src/main/java/br/com/saborearte/model/ReceitaIngrediente.java
@@ -1,12 +1,14 @@
 package br.com.saborearte.model;
 
+import java.math.BigDecimal;
+
 public class ReceitaIngrediente {
 
     // ===== Atributos =====
 
     private int id_receita;
     private int id_ingrediente;
-    private int quantidade_receita_ingrediente;
+    private BigDecimal quantidade_receita_ingrediente;
     private String unidade_medida_receita_ingrediente;
 
     // ===== Campo extra (NÃO existe na tabela receita_ingrediente — não persistir) =====
@@ -24,7 +26,7 @@ public class ReceitaIngrediente {
 
     public ReceitaIngrediente(int id_receita,
                                int id_ingrediente,
-                               int quantidade_receita_ingrediente,
+                               BigDecimal quantidade_receita_ingrediente,
                                String unidade_medida_receita_ingrediente) {
 
         this.id_receita = id_receita;
@@ -41,8 +43,8 @@ public class ReceitaIngrediente {
     public int getId_ingrediente() { return id_ingrediente; }
     public void setId_ingrediente(int id_ingrediente) { this.id_ingrediente = id_ingrediente; }
 
-    public int getQuantidade_receita_ingrediente() { return quantidade_receita_ingrediente; }
-    public void setQuantidade_receita_ingrediente(int quantidade_receita_ingrediente) { this.quantidade_receita_ingrediente = quantidade_receita_ingrediente; }
+    public BigDecimal getQuantidade_receita_ingrediente() { return quantidade_receita_ingrediente; }
+    public void setQuantidade_receita_ingrediente(BigDecimal quantidade_receita_ingrediente) { this.quantidade_receita_ingrediente = quantidade_receita_ingrediente; }
 
     public String getUnidade_medida_receita_ingrediente() { return unidade_medida_receita_ingrediente; }
     public void setUnidade_medida_receita_ingrediente(String unidade_medida_receita_ingrediente) { this.unidade_medida_receita_ingrediente = unidade_medida_receita_ingrediente; }
